+++ src/main/java/module-info.java
@@ -1,5 +1,5 @@
 // ChangeMe: to match the module, i.e. `creek.<repo-name>.<module-name>`:
 // Also search for references to `reek.template.module.single` and replace accordingly.
 module creek.template.module.single {
-    exports org.creek.api.example;
+    exports org.creekservice.api.example;
 }
